Revert change to tests since python venv not in bin/omero
